Change create to drop schema and stop migrations
Now we use the script that was used to also create the schema first to drop it instead. Being outside of the db-migrate it doesn't try to record the action.

We update the migration command (and drop the others whilst at it) to now only run our `drop-schema.js` script.