Fix dropping schema solution (#1112)
DEFRA/water-abstraction-team#143

> See our Team issue for an expanded explanation

In [Remove charge versions import process](https://github.com/DEFRA/water-abstraction-import/pull/1111), we realised that having removed the process, the `water_import` schema was redundant. We can finally drop another schema!

We created a migration to do just that, and all ran fine locally. We thought our issues in CI were specific to that, but when we tried to deploy to our non-production environments, we got the same error.

```
> node scripts/create-schema && node scripts/migrate

OK
ERROR: Command failed: export DATABASE_URL=postgres://waterdbu01:waterdbu01@wal-rds-pre.aws-int.defra.cloud:5432/water-abstraction-preprod && db-migrate up
[ERROR] 20250701144437-drop-water-import-schema error: relation "migrations" does not exist
    at Parser.parseErrorMessage (/srv/app/node/water-abstraction-import/releases/20250701_212348/node_modules/pg-protocol/dist/parser.js:285:98)
    at Parser.handlePacket (/srv/app/node/water-abstraction-import/releases/20250701_212348/node_modules/pg-protocol/dist/parser.js:122:29)
    at Parser.parse (/srv/app/node/water-abstraction-import/releases/20250701_212348/node_modules/pg-protocol/dist/parser.js:35:38)
    at Socket.<anonymous> (/srv/app/node/water-abstraction-import/releases/20250701_212348/node_modules/pg-protocol/dist/index.js:11:42)
    at Socket.emit (node:events:518:28)
    at addChunk (node:internal/streams/readable:561:12)
    at readableAddChunkPushByteMode (node:internal/streams/readable:512:3)
    at Readable.push (node:internal/streams/readable:392:5)
    at TCP.onStreamRead (node:internal/stream_base_commons:189:23) {
  length: 109,
  severity: 'ERROR',
  code: '42P01',
  detail: undefined,
  hint: undefined,
  position: '13',
  internalPosition: undefined,
  internalQuery: undefined,
  where: undefined,
  schema: undefined,
  table: undefined,
  column: undefined,
  dataType: undefined,
  constraint: undefined,
  file: 'parse_relation.c',
  line: '1395',
  routine: 'parserOpenTable'
}
[ERROR] AssertionError [ERR_ASSERTION]: ifError got unwanted exception: relation "migrations" does not exist
```

It appears that dropping the schema has caused the migration to fail because the table it is trying to update is located in the `water_import` schema.

This means we cannot use [db-migrate](https://db-migrate.readthedocs.io/en/latest/) and the migration process to drop the schema.

Instead, we replace the `create-schema.js` script that is called before db-migrate with a `drop-schema.js`, and no longer call db-migrate.

And since we are no longer running migrations, we can take this opportunity to eliminate them!