Drop extract old lines import process (#1113)
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the 'extract-old-lines' process we added to allow us to import NALD return lines in preparation for WRLS taking control of returns.

Since we switched from NALD to WRLS on [June 11](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.35.1), this process has been disabled and now will never be re-enabled. This means we can remove it altogether.

The same can be said for the table we created when first importing the old lines from the extract. We only needed it the first time we started importing NALD return submission data on [May 9](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.34.0). It's a large table that's no longer in use, so we should remove it from all environments.