Remove return versions import process
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the 'return-versions-import' process. The legacy import was already importing some of this data. But to support WRLS taking over from NALD we re-wrote and expanded what the legacy import was doing.

Since we switched from NALD to WRLS on [June 11](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.35.1), this process has been disabled and now will never be re-enabled. This means we can remove it altogether.