Remove licence submissions import process
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the 'licence-submissions-import' process. The legacy import had an ability to import returns submissions data. But it was only intended for test environments. We fixed the issues with it, and re-used it to handle importing _all_ the NALD submission data missing from WRLS.

Since we switched from NALD to WRLS on [June 11](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.35.1), this process has been disabled and now will never be re-enabled. This means we can remove it altogether.