Remove invalid returns cleanup process
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the 'invalid-returns-cleanup' process.As we started importing returns data from NALD, we found the NALD processes allowed for data to become corrupted. So, as part of the import we added a process that fixed/corrected/deleted this invalid data.

Since we switched from NALD to WRLS on [June 11](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.35.1), this process has been disabled and now will never be re-enabled. This means we can remove it altogether.