Remove invalid returns cleanup process (#1118)
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the 'invalid-returns-cleanup' process. As we began importing return data from NALD, we discovered that the NALD processes allowed data to become corrupted. As part of the import, we added a process that corrected and deleted this invalid data.

Since we switched from NALD to WRLS on [June 11](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.35.1), this process has been disabled and now will never be re-enabled. This means we can remove it altogether.