Fix licence clean process
https://eaflood.atlassian.net/browse/WATER-4662

As part of preparing for WRLS taking over returns management from NALD we needed to ensure the return version information from NALD was being imported.

When we dug in we realised there was data in WRLS that no longer existed in NALD. This is because NALD, unlike WRLS, allows users to delete records. We cleaned all that up (see WATER-4654 ) but that made us look at licence data where we found the same problem.

@Jozzey did a great job of adding a series of queries that would [clean up deleted licence data](https://github.com/DEFRA/water-abstraction-import/pull/1047) in WRLS _safely_ (we can't delete stuff that is linked to a bill run for example).

Unfortunately, @Cruikshanks appears to have butchered it when he [Refactored all overnight jobs into one](https://github.com/DEFRA/water-abstraction-import/pull/1063). So now were deleting most records, just not the licences themselves! 🤦

This change reviews what we have now with what @Jozzey originally did and ensures we are deleting everything we're supposed to be.