Spotted the problem
The original 'job' had a `cleanDeletedLicences()` function that is mirrored in our `deleted-licences.js` module.

It fired a number of queries (all those in our module) as part of a `Promise.all()` call.

What we failed to spot was it then fired some additional queries one at a time _after_ that call. It's these that we are missing.