Update cleanup process to delete 'draft' licences
https://eaflood.atlassian.net/browse/WATER-5247

The legacy import process imported all licences found in import.NALD_ABS_LICENCES.

When we overhauled the process in water-abstraction-import we corrected that to only be licences that were not in a 'DRAFT' status.

The legacy version also never deleted anything. So, once a record was imported into WRLS, if a user then deleted it in NALD, the record would remain in WRLS.

Again, the overhauled version added a 'clean' process that would delete licence data from WRLS that no longer exists in NALD.

We've spotted licences in WRLS production have a status of DRAFT in NALD.

In theory, they should have been deleted as part of the clean. But there is a flaw in the query that the clean code runs.

- Import works from a list of all licences that are not in DRAFT
- Clean simply checks if the licence exists in NALD

So, the query in clean isn't returning these for deletion because it sees they exist. However, we're not importing any new draft licenses due to the change in the import.

In this change we update the clean query to use the same query the import uses to determine what is a 'live' licence in NALD.

Then the clean will include these four licences in its clean up.