Correct and delete crmV2 document roles
https://eaflood.atlassian.net/browse/WATER-4891
https://eaflood.atlassian.net/browse/WATER-5227

We had [another report of licence holder details being wrong](https://eaflood.atlassian.net/browse/WATER-5227) in WRLS.

When we looked it's because the data in `crm_v2.document_roles` is out of sync with that in the NALD import.

We've known for some time now that part of the problem is that when a licence version is deleted in NALD, we don't delete the relevant record in `crm_v2.document_roles`.

We now have the 'clean' process for all other licence data, but because `crm_v2` we though was intended as a CRM solution, we held off deleting anything. We also faced a lack of understanding. On the face of it, there should be a one-to-one relationship with licence version and `crm_v2.document_role`.

However, the legacy import seemed to be trying to identify replaced licence versions, so the end result did not always line up as one-to-one.

Having had another incident of incorrect licence holder causing a problem, we figured it was time to get to the bottom of what is going on.

Fortunately, WATER-5227 reported an issue with two licences and the issue for each we think have helped us figure out what is going wrong.

**Deleted or changed licence versions**

One of the licences was displaying the issue we were aware of. In the UI we are displaying 3 licence holder records for the licence, of which only one is correct. When a user is attempting to create a billing account, we're displaying the wrong details because the query is picking the first result.

In the database we can see there are actually 4 licence holder records. Having checked the NALD data, we've confirmed there are only two versions. What appears to have happened is that the start dates for the licence versions have been corrected. This has resulted in the import creating two records new `crm_v2.document_roles` records, hence there are now four.

To further complicate matters, the licence has an expiry date set to 2039. We don't know when this was applied to the NALD licence versions, but what we now have in WRLS is three records with that expiry date. This is why the UI is returning 3 records rather than 1 when displaying the current licence holder.

**Incorrect sorting**

The second licence has four licence versions in NALD, one of which is an increment that has the same start date as its predecessor. In WRLS we only have four `crm_v2.document_roles`. This is the confusion mentioned earlier. WRLS is not trying to create a one-to-one mapping. Instead, if two licence versions have the same start date, it will import the details for the latest issue or increment, not both.

Or that's what it was supposed to do! After digging into this one, we've found that the sort logic is wrong. It was sorting by issue numbers in ascending order, but then when an increment was encountered it was sorting in descending order! This mix-up we believe has led to `crm_v2.document_roles` being setup incorrectly. In this case it means the the end date is not being applied, resulting in two licence holder records without end dates, hence both are displayed in the UI as the 'current' licence holder.

**Unnecessary replacement filtering**

The intent of the sorting was to 'filter' out replaced licence versions. The legacy logic would first identify if two licence versions have the same start date. If they did, it would then sort the licence versions so it could 'filter' out the replaced licence version. But the fault in the sort meant it was actually importing the replaced version!

Stepping back though, the concept of filtering out replaced licence versions just overcomplicates the process. If the licence versions are sorted in ascending order correctly, they will then be imported in that order. Where a licence version replaces its predecessor, it does not matter that we import both. For example, imagine we have the following NALD licence versions

- Issue 1, increment 1, start date 2024-01-01, licence holder 'Barney Rubble'
- Issue 1, increment 2, start date 2024-01-01, licence holder 'Fred Flintstone'

Because we are using the `ON CONFLICT (document_id, role_id, start_date)` constraint in the UPSERT query, the import will simply overwrite the WRLS record with the latest licence version details, in this case setting it to 'Fred Flintstone'.

Yes, it could be argued that we are running a query against the DB needlessly. But the process becomes much simpler to follow, and less likely to result in an issue.

---

So, this change does 3 things

- It drops the logic around attempting to filter out replacements
- It orders _licence versions_ (not just when two have the same start date) in ascending order correctly then imports them
- It deletes any `crm_v2.document_roles` records which were not updated in the import

With these 3 changes we should prevent future instances of licence holder details being incorrect.