Stop importing disabled or ended roles (contacts)
https://eaflood.atlassian.net/browse/WATER-5374

This change was initiated by an issue with the returns invitations sent for the summer 2024/25 cycle (sent on 3 November).

The Billing & Data team spotted that invitations were sent to contacts that are marked as ended in NALD.

NALD captures an effective start and effective end date against a licence's 'roles'. They spotted that the new notice engine had sent invites to roles event though 3 November 2025 was greater than their 'effective end date'.

When we built the notice engine we had two pots of CRM data to choose from.

- The contacts held in the metadata of `crm.document_header`
- The contacts held in `crm_v2.document_roles`

Our review of the legacy code told us it was using `crm.document_header`, plus we were told at the time that this information was more accurate than what folks saw in `crm_v2.document_roles`.

The problem is unlike `crm_v2.document_roles`, we don't have these effective start and end dates in the data nor do they impact the import; everything is imported.

Now we know that NALD roles have start and end dates, plus a 'disabled' flag, and the intent of contacts `crm.document_header` is that they are the 'current' contacts, we can solve the problem with the notice engine at source.

This change updates the `licence-crm-import` step to ignore

- Roles that have not started
- Roles that have ended
- Roles that are disabled