Add shai-hulud-detect as a daily workflow (#1137)
https://eaflood.atlassian.net/browse/WATER-5406

With the [announcement that the npm supply-chain Shai-Hulud attack is back](https://www.koi.ai/incident/live-updates-sha1-hulud-the-second-coming-hundred-npm-packages-compromised), we're being asked yet again to confirm if our services have been compromised.

The problem is, as a self-replicating worm, which npm packages are compromised is a moving target.

Some of our fantastic colleagues have pulled together bash scripts that can be run to test whether a project may have been compromised. Initially, they depended on self-maintained CSV files of compromised packages, but later iterations were pulling down [compromised-packages.txt](https://github.com/Cobenian/shai-hulud-detect/blob/main/compromised-packages.txt).

We'd already found [shai-hulud-detect](https://github.com/Cobenian/compromised-packages.txt) when looking for something 'off-the-shelf', so figured if we're using their list, why not also use their scanner?

We initially considered including the scan as a CI step. But we have seen that its scanning can change depending on whether `npm ci` is run before or after the scan.

Before, the scan was pretty quick. After, it can take more than an hour as a workflow action. The difference is the depth and types of scanning being performed.

So, we've opted to isolate this as a separate workflow, which we'll schedule daily, whilst we continue to monitor and develop how we check the service.

This first iteration uses the 'simple' version, i.e., we don't run 'npm ci' before the scan. There are issues with 'false' positives in the more thorough scan, which [we are continuing to monitor](https://github.com/DEFRA/water-abstraction-permit-repository/pull/398).

* Add workflow

* Add to README