Add permissions to scanner workflow
https://eaflood.atlassian.net/browse/WATER-5406

After adding a new workflow to scan for compromised npm packages with [shai-hulud-detect](https://github.com/Cobenian/shai-hulud-detect)  to _all_ our repos, GitHub suddenly piped up in one of them that it really should declare permissions.

As we are only cloning the code from the repo, we only need to declare `permissions: read`. It just would have been nice for it to tell us before we merged them all!