Unflag as 'deleted' restored document headers
https://eaflood.atlassian.net/browse/WATER-5433

It was reported that an external user could no longer access a licence on the external site. When checked on the internal site we can see the licence is live and registered to them.

The link from user to licence is not the clearest. In the `idm.users` record, if an external user is linked to a licence the `external_id` will be populated. This is actually the `entity_id` from `crm.entity`. When you check you'll find a record like the following example.

|entity_id|entity_nm             |entity_type|
|---------|----------------------|-----------|
|e001     |external@example.co.uk|individual |

`crm.entity_roles` will tell you whether they are the 'primary user' or a 'returns agent'. But it is also the link to the licence.

|entity_role_id|entity_id|role        |company_entity_id|
|--------------|---------|------------|-----------------|
|er001         |e001     |primary_user|ce001            |

Any `crm.document_header` record linked to `crm.entity_roles` via `company_entity_id` is a 'licence' registered to the user.

|document_header_id|company_entity_id|system_external_id|date_deleted|
|------------------|-----------------|------------------|------------|
|dh001             |ce001            |01/123            |            |
|dh002             |ce001            |01/123/R01        |2025-06-26  |

> Don't shout at us! None of this makes sense! 😩😭

In the case of the external user, the licence they couldn't see had been flagged as deleted, just like in our example above.

`crm.document_header` records are flagged as deleted by the **flag-deleted-documents** step in the import process. It checks for any `crm.document_header` records which have a `system_external_id` (licence reference) that no longer exists in `import.NALD_ABS_LICENCES`. It's the same way we identify licence and licence-related date that needs to be deleted in **clean**.

Our assumption is that the licence was deleted from NALD, and then reinstated. The external site is legacy code which depends heavily on `crm.document_header` and `permit.licence`, schemas and tables we no longer use in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). This explains why they cannot see it, but we can internally.

We could have run a 'one-off' migration to fix it. But we found another example when we checked the data, so it feasibly could happen again. Therefore, this change updates the **flag-deleted-documents** step to reinstate 'documents' (licences) that are no longer deleted in NALD.